requirements.txt
discord.py
gitpython
python-dotenv